+++ docs/changelog.rst
@@ -26,6 +26,7 @@ Deprecations:
 Changes:
 ^^^^^^^^
 
+-
 
 
 ----