+++ docs/changelog.rst
@@ -28,8 +28,8 @@ Deprecations:
 Changes:
 ^^^^^^^^
 
-- Attributes set by :func:`characteristic.attributes` are now stored on the
-  class as well.
+- Attributes set by :func:`characteristic.attributes` are now stored on the class as well.
+  [`20 <https://github.com/hynek/characteristic/pull/20>`_]
 
 
 ----